UP: docs + edits

+++ mobile/app/src/main/java/com/example/test/Cash.kt
@@ -1,4 +1,4 @@
 package com.example.test
 
-class Cash(val accountNumber: String, val accountName: String, var amount: Int?) {
+class Cash(val accountNumber: String, val accountName: String, var amount: Float?) {
 }
